+++ client/src/pages/Login.js
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { useMutation } from '@apollo/client';
 import { Form, Button, Alert } from 'react-bootstrap';
-import { ADD_USER } from '../../utils/mutations';
-import Auth from '../../utils/auth';
+import { ADD_USER } from '../utils/mutations';
+import Auth from '../utils/auth';
 
 const Login = () => {
